-vpic_base; psc, cmake: mv vpic_base_init
