+++ src/libpsc/psc.cxx
@@ -78,7 +78,45 @@ _psc_read(struct psc *psc, struct mrc_io *io)
 #endif
 
 // FIXME
-void vpic_base_init(int* pargc, char*** pargv);
+void vpic_base_init(int* pargc, char*** pargv)
+{
+  static bool vpic_base_inited = false;
+
+  if (vpic_base_inited) {
+    return;
+  }
+  vpic_base_inited = true;
+
+  //  boot_services( &argc, &argv );
+  {
+    // Start up the checkpointing service.  This should be first.
+#ifdef USE_VPIC
+    boot_checkpt(pargc, pargv);
+
+    serial.boot(pargc, pargv);
+    thread.boot(pargc, pargv);
+
+    // Boot up the communications layer
+    // See note above about thread-core-affinity
+
+    boot_mp(pargc, pargv);
+#else
+    int provided;
+    MPI_Init_thread(pargc, pargv, MPI_THREAD_MULTIPLE, &provided);
+#endif
+
+    MPI_Comm_dup(MPI_COMM_WORLD, &psc_comm_world);
+    MPI_Comm_rank(psc_comm_world, &psc_world_rank);
+    MPI_Comm_size(psc_comm_world, &psc_world_size);
+
+    MPI_Barrier(psc_comm_world);
+#ifdef USE_VPIC
+    _boot_timestamp = 0;
+    _boot_timestamp = uptime();
+#endif
+  }
+  LOG_INFO("vpic_base_init() done\n");
+}
 
 void psc_init(int& argc, char**& argv)
 {
